fix: 모델파일명 변경
label->tag, labellist->label db테이블명을 적절하게 바꿈으로써 파일명들도 변경
