--- BE/src/interfaces/labellist.ts
@@ -1,8 +1,0 @@
-/* eslint-disable camelcase */
-export interface LabelList {
-  id: number | null;
-  name: string;
-  description: string;
-  color: string;
-  create_at: Date;
-}
