+++ BE/src/interfaces/tag.ts
@@ -1,0 +1,6 @@
+/* eslint-disable camelcase */
+export interface Label {
+  id: number | null;
+  issue_id: number;
+  iabel_id: number;
+}
