--- BE/src/models/label.ts
@@ -1,5 +1,0 @@
-import db from "../providers/database";
-import Model from "./model";
-import { Label } from "../interfaces/label";
-
-export default class LabelModel extends Model {}
